+++ nautilus_trader/cache/cache.pyx
@@ -1689,7 +1689,12 @@ cdef class Cache(CacheFacade):
 
         self._log.debug(f"Snapshot {repr(copied_position)}.")
 
-    cpdef void snapshot_position_state(self, Position position, uint64_t ts_snapshot):
+    cpdef void snapshot_position_state(
+        self,
+        Position position,
+        uint64_t ts_snapshot,
+        bint open_only=True,
+    ):
         """
         Snapshot the state dictionary for the given `position`.
 
@@ -1701,9 +1706,15 @@ cdef class Cache(CacheFacade):
             The position to snapshot the state for.
         ts_snapshot : uint64_t
             The UNIX timestamp (nanoseconds) when the snapshot was taken.
+        open_only : bool, default True
+            If only open positions should be snapshot, this flag helps to avoid race conditions
+            where a position is snapshot when no longer open.
 
         """
         Condition.not_none(position, "position")
+
+        if open_only and not position.is_open_c():
+            return  # Only snapshot open positions
 
         if self._database is None:
             self._log.warning(