+++ nautilus_trader/live/execution_client.py
@@ -21,6 +21,7 @@ from __future__ import annotations
 
 import asyncio
 import functools
+import traceback
 from asyncio import Task
 from collections.abc import Coroutine
 from datetime import timedelta
@@ -36,6 +37,7 @@ from nautilus_trader.common.providers import InstrumentProvider
 from nautilus_trader.core.correctness import PyCondition
 from nautilus_trader.core.uuid import UUID4
 from nautilus_trader.execution.client import ExecutionClient
+from nautilus_trader.execution.messages import BatchCancelOrders
 from nautilus_trader.execution.messages import CancelAllOrders
 from nautilus_trader.execution.messages import CancelOrder
 from nautilus_trader.execution.messages import ModifyOrder
@@ -199,18 +201,21 @@ class LiveExecutionClient(ExecutionClient):
         success: str | None,
         task: Task,
     ) -> None:
-        if task.exception():
+        e: BaseException | None = task.exception()
+        if e:
+            tb_str = "".join(traceback.format_exception(type(e), e, e.__traceback__))
             self._log.error(
-                f"Error on `{task.get_name()}`: " f"{task.exception()!r}",
+                f"Error on `{task.get_name()}`: " f"{task.exception()!r}\n{tb_str}",
             )
         else:
             if actions:
                 try:
                     actions()
                 except Exception as e:
+                    tb_str = "".join(traceback.format_exception(type(e), e, e.__traceback__))
                     self._log.error(
                         f"Failed triggering action {actions.__name__} on `{task.get_name()}`: "
-                        f"{e!r}",
+                        f"{e!r}\n{tb_str}",
                     )
             if success:
                 self._log.info(success, LogColor.GREEN)
@@ -265,6 +270,12 @@ class LiveExecutionClient(ExecutionClient):
         self.create_task(
             self._cancel_all_orders(command),
             log_msg=f"cancel_all_orders: {command}",
+        )
+
+    def batch_cancel_orders(self, command: BatchCancelOrders) -> None:
+        self.create_task(
+            self._batch_cancel_orders(command),
+            log_msg=f"batch_cancel_orders: {command}",
         )
 
     def query_order(self, command: QueryOrder) -> None:
@@ -494,4 +505,9 @@ class LiveExecutionClient(ExecutionClient):
     async def _cancel_all_orders(self, command: CancelAllOrders) -> None:
         raise NotImplementedError(  # pragma: no cover
             "implement the `_cancel_all_orders` coroutine",  # pragma: no cover
+        )
+
+    async def _batch_cancel_orders(self, command: BatchCancelOrders) -> None:
+        raise NotImplementedError(  # pragma: no cover
+            "implement the `_batch_cancel_orders` coroutine",  # pragma: no cover
         )