+++ RELEASES.md
@@ -7,6 +7,7 @@ Released on TBD (UTC).
 - Added `Cache.is_order_pending_cancel_local(...)` (tracks local orders in cancel transition)
 - Added `BinanceTimeInForce.GTD` enum member (futures only)
 - Added package version check for `nautilus_ibapi`, thanks @rsmb7z
+- Added `RiskEngine` min/max instrument notional limit checks
 
 ### Breaking Changes
 None
@@ -16,6 +17,8 @@ None
 - Fixed `OrderEmulator` start-up processing of OTO contingent orders (when position from parent is open)
 - Fixed `SandboxExecutionClientConfig` `kw_only=True` to allow importing without initializing
 - Fixed `OrderBook` pickling (did not include all attributes), thanks @limx0
+- Fixed `Binance` instruments missing max notional values, thanks for reporting @AnthonyVince and thanks for fixing @filipmacek
+- Fixed open position snapshots race condition (added `open_only` flag)
 
 ---